style: 스타일 분리
