+++ src/components/Admin/PenaltyAdd.tsx
@@ -1,68 +1,10 @@
-import theme from '@/styles/theme';
-import { styled } from 'styled-components';
-import { Wrapper } from '@/styles/admin/DuesRegisterDropDown.styled';
 import Button from '@/components/Admin/Button';
-import * as S from '@/styles/admin/DuesRegisterDropDown.styled';
+import * as S from '@/styles/admin/penalty/PenaltyAdd.styled';
+import { ButtonWrapper } from '@/styles/admin/DuesRegisterDropDown.styled';
 import { useState } from 'react';
 import { useMemberContext } from '@/components/Admin/context/MemberContext';
 import PenaltyMemberDropdown from '@/components/Admin/PenaltyMemberDropdown';
 import { postPenaltyApi } from '@/api/admin/penalty/getPenalty';
-
-export const TitleWrapper = styled.div`
-  padding: 25px 30px;
-`;
-export const SubTitle = styled.div`
-  font-family: ${theme.font.regular};
-  font-size: 16px;
-  margin-left: 10px;
-  margin-bottom: 10px;
-`;
-
-export const PenaltyWrapper = styled(Wrapper)`
-  border-radius: 4px;
-  width: 490px;
-  min-width: 490px;
-  min-height: 420px;
-  margin-top: 30px;
-  border-top: 0px;
-  padding: 0px;
-`;
-
-export const InputWrapper = styled.div`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  margin-bottom: 15px;
-`;
-export const Input = styled.input`
-  width: 100%;
-  font-size: 18px;
-  font-family: ${theme.font.regular};
-  border: 1px solid #dedede;
-  border-radius: 4px;
-  padding: 12px 16px;
-  box-sizing: border-box;
-  &:focus {
-    outline: 1.5px solid ${theme.color.gray[18]};
-  }
-`;
-
-export const Title = styled.div`
-  font-size: 20px;
-`;
-
-export const Line = styled.div`
-  border: 1px solid #dedede;
-`;
-
-export const ItemWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 30px;
-  padding: 30px;
-`;
 
 const PenaltyAdd: React.FC = () => {
   const { members } = useMemberContext();
@@ -120,15 +62,15 @@ const PenaltyAdd: React.FC = () => {
     }
   };
   return (
-    <PenaltyWrapper>
-      <TitleWrapper>
-        <Title>패널티 추가</Title>
-      </TitleWrapper>
-      <Line />
-      <ItemWrapper>
-        <InputWrapper>
-          <SubTitle>이름</SubTitle>
-          <Input
+    <S.PenaltyWrapper>
+      <S.TitleWrapper>
+        <S.Title>패널티 추가</S.Title>
+      </S.TitleWrapper>
+      <S.Line />
+      <S.ItemWrapper>
+        <S.InputWrapper>
+          <S.SubTitle>이름</S.SubTitle>
+          <S.Input
             placeholder="이름"
             value={searchTerm}
             onChange={(e) => {
@@ -144,16 +86,16 @@ const PenaltyAdd: React.FC = () => {
               onSelect={handleSelectMember}
             />
           )}
-        </InputWrapper>
-        <InputWrapper>
-          <SubTitle>패널티 사유</SubTitle>
-          <Input
+        </S.InputWrapper>
+        <S.InputWrapper>
+          <S.SubTitle>패널티 사유</S.SubTitle>
+          <S.Input
             placeholder="ex) 미션 과제 미제출"
             value={penaltyDescription}
             onChange={(e) => setPenaltyDescription(e.target.value)}
           />
-        </InputWrapper>
-        <S.ButtonWrapper>
+        </S.InputWrapper>
+        <ButtonWrapper>
           <Button
             description="초기화"
             color="#323232"
@@ -168,9 +110,9 @@ const PenaltyAdd: React.FC = () => {
             borderRadius="4px"
             onClick={handleAddPenalty}
           />
-        </S.ButtonWrapper>
-      </ItemWrapper>
-    </PenaltyWrapper>
+        </ButtonWrapper>
+      </S.ItemWrapper>
+    </S.PenaltyWrapper>
   );
 };
